+++ lib/core/localisations/language_collection/languageEn.dart
@@ -27,6 +27,9 @@ class LanguageEn extends Languages {
   @override
   String get va => "Voice Assistant";
 
+  @override
+  String get anim => "Animations";
+
   ///home
   @override
   String get homeLabel => "You have pushed the button this many times:";